Enhanced security and token cleaning

+++ src/Controller/SecurityController.php
@@ -8,6 +8,7 @@ use App\Entity\User;
 use App\Form\UserType;
 use App\Repository\UserRepository;
 use App\Repository\ResetPasswordRepository;
+use Symfony\Component\RateLimiter\RateLimiterFactory;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bridge\Twig\Mime\TemplatedEmail;
 use App\Security\LoginFormAuthenticator;
@@ -75,9 +76,15 @@ class SecurityController extends AbstractController
 
 
     #[Route('/reset-password/{token}', name: 'reset-password')]
-      public function resetPassword(UserPasswordHasherInterface $userPasswordHasher, Request $request, EntityManagerInterface $em, string $token, ResetPasswordRepository $resetPasswordRepository)
+      public function resetPassword(RateLimiterFactory $passwordRecoveryLimiter, UserPasswordHasherInterface $userPasswordHasher, Request $request, EntityManagerInterface $em, string $token, ResetPasswordRepository $resetPasswordRepository)
       {
-        $resetPassword = $resetPasswordRepository->findOneBy(['token' => $token]);
+        $limiter = $passwordRecoveryLimiter->create($request->getClientIp());
+        if (false === $limiter->consume(1)->isAccepted()) {
+          $this->addFlash('error', 'Vous devez attendre 1 heure pour refaire une tentative');
+          return $this->redirectToRoute('login');
+        }
+
+        $resetPassword = $resetPasswordRepository->findOneBy(['token' => sha1($token)]);
         if (!$resetPassword || $resetPassword->getExpiredAt() < new \DateTime('now')) {
           if ($resetPassword) {
             $em->remove($resetPassword);
@@ -121,8 +128,14 @@ class SecurityController extends AbstractController
 
 
     #[Route('/reset-password-request', name: 'reset-password-request')]
-    public function resetPasswordRequest(MailerInterface $mailer, Request $request, UserRepository $userRepository, ResetPasswordRepository $resetPasswordRepository, EntityManagerInterface $em)
+    public function resetPasswordRequest(RateLimiterFactory $passwordRecoveryLimiter, MailerInterface $mailer, Request $request, UserRepository $userRepository, ResetPasswordRepository $resetPasswordRepository, EntityManagerInterface $em)
     {
+        $limiter = $passwordRecoveryLimiter->create($request->getClientIp());
+        if (false === $limiter->consume(1)->isAccepted()) {
+          $this->addFlash('error', 'Vous devez attendre 1 heure pour refaire une tentative');
+          return $this->redirectToRoute('login');
+        }
+
         $emailForm = $this->createFormBuilder()->add('email', EmailType::class, [
             'constraints' => [
               new NotBlank([
@@ -145,7 +158,7 @@ class SecurityController extends AbstractController
                 $resetPassword->setUser($user);
                 $resetPassword->setExpiredAt(new \DateTimeImmutable('+2 hours'));
                 $token = substr(str_replace(['+', '/', '='], '', base64_encode(random_bytes(30))), 0, 20);
-                $resetPassword->setToken($token);
+                $resetPassword->setToken(sha1($token));
                 $em->persist($resetPassword);
                 $em->flush();
                 $email = new TemplatedEmail();
